chore: implement new bottom view component
