+++ demo/App.tsx
@@ -1,9 +1,9 @@
 import './src/gesture-handler';
 import * as Linking from 'expo-linking';
-import { BottomTabBar, BottomTabView } from './src/Core';
 import { Logger, fetchWrapper, formatDate } from './src/Helpers';
 import Behaviors from './src/Behaviors';
 import { BottomTabBarContextProvider } from './src/Contexts';
+import { BottomTabView } from './src/Core';
 import Components from './src/Components';
 import Constants from 'expo-constants';
 import ElementErrorComponent from './src/element-error-component';
@@ -50,10 +50,7 @@ export default () => (
           formatDate={formatDate}
           loadingScreen={LoadingScreen}
           logger={new Logger(Logger.Level.log)}
-          navigationComponents={{
-            BottomTabBar,
-            BottomTabView,
-          }}
+          navigationComponents={{ BottomTabView }}
         />
       </BottomTabBarContextProvider>
     </NavigationContainer>
