+++ demo/src/Core/BottomTabBar/BottomTabView.tsx
@@ -1,38 +1,117 @@
+import * as React from 'react';
+import type { BottomTabViewProps, NativeTabEvent } from './types';
 import { getNativeTabItems, isLiquidGlassSupported } from './helpers';
-import type { BottomTabViewProps } from './types';
-import NativeTabView from './NativeTabView';
+import { BottomTabBar } from '.';
+import type { BottomTabBarProps } from '@react-navigation/bottom-tabs';
+import { CommonActions } from '@react-navigation/native';
+import type { NativeSyntheticEvent } from 'react-native';
 import { BottomTabView as RNBottomTabView } from '@react-navigation/bottom-tabs';
-import React from 'react';
+import { Tabs } from 'react-native-screens';
 import { useBottomTabBarContext } from '../../Contexts';
 
 export const BottomTabView = (props: BottomTabViewProps) => {
   const { getElementProps } = useBottomTabBarContext();
+  const [confirmedProvenance, setConfirmedProvenance] = React.useState(0);
   const navigatorId = props.navigation.getId?.();
-  const items =
-    isLiquidGlassSupported() && getElementProps && navigatorId
-      ? getNativeTabItems(getElementProps(navigatorId))
+  const nativeTabItems =
+    isLiquidGlassSupported() && navigatorId
+      ? getNativeTabItems(getElementProps?.(navigatorId))
       : false;
 
-  if (items) {
+  if (!nativeTabItems) {
+    const tabBar = (tabBarProps: BottomTabBarProps) => {
+      if (nativeTabItems === undefined || !navigatorId) {
+        return null;
+      }
+      return (
+        <BottomTabBar
+          descriptors={tabBarProps.descriptors}
+          id={navigatorId}
+          insets={tabBarProps.insets}
+          navigation={tabBarProps.navigation}
+          state={tabBarProps.state}
+        />
+      );
+    };
+
     return (
-      <NativeTabView
+      <RNBottomTabView
         descriptors={props.descriptors}
-        items={items}
         navigation={props.navigation}
         state={props.state}
+        tabBar={tabBar}
       />
     );
   }
 
+  const focusedRouteKey = props.state.routes[props.state.index].key;
+  const onTabSelectionChange = (
+    event: NativeSyntheticEvent<NativeTabEvent>,
+  ) => {
+    const { actionOrigin, provenance, selectedScreenKey } = event.nativeEvent;
+    const route = props.state.routes.find(
+      item => item.key === selectedScreenKey,
+    );
+
+    if (!route) {
+      return;
+    }
+
+    const tabPressEvent =
+      actionOrigin === 'user'
+        ? props.navigation.emit({
+            canPreventDefault: true,
+            target: route.key,
+            type: 'tabPress',
+          })
+        : undefined;
+
+    setConfirmedProvenance(provenance);
+    if (
+      tabPressEvent?.defaultPrevented ||
+      actionOrigin === 'programmatic-js' ||
+      focusedRouteKey === route.key
+    ) {
+      return;
+    }
+
+    props.navigation.dispatch({
+      ...CommonActions.navigate({ merge: true, name: route.name }),
+      target: props.state.key,
+    });
+  };
+
   return (
-    <RNBottomTabView
-      descriptors={props.descriptors}
-      detachInactiveScreens={props.detachInactiveScreens}
-      navigation={props.navigation}
-      safeAreaInsets={props.safeAreaInsets}
-      sceneContainerStyle={props.sceneContainerStyle}
-      state={props.state}
-      tabBar={items === undefined ? () => null : props.tabBar}
-    />
+    <Tabs.Host
+      ios={{ tabBarMinimizeBehavior: 'onScrollDown' }}
+      navStateRequest={{
+        baseProvenance: confirmedProvenance,
+        selectedScreenKey: focusedRouteKey,
+      }}
+      onTabSelected={onTabSelectionChange}
+      onTabSelectionRejected={onTabSelectionChange}
+      rejectStaleNavStateUpdates
+    >
+      {nativeTabItems.map(item => {
+        const route = props.state.routes.find(
+          ({ name }) => name === item.route,
+        );
+        if (!route) {
+          return null;
+        }
+
+        return (
+          <Tabs.Screen
+            key={route.key}
+            badgeValue={item.badgeValue}
+            ios={{ icon: item.icon }}
+            screenKey={route.key}
+            title={item.label ?? route.name}
+          >
+            {props.descriptors[route.key].render()}
+          </Tabs.Screen>
+        );
+      })}
+    </Tabs.Host>
   );
 };
